fix: train_mi consistency, workflow merge, dimensionality parallelisation
Five coordinated fixes across the library:

1. **Restore all_mi_negative guard (run.py estimate mode)**
   The explicit zero guard is preserved: if all test-MI values during
   training were non-positive the estimate is set to 0.0 (trainer already
   sets train_mi=0.0, but the guard makes the reason explicit in code).

2. **precision.py → train data & train_mi throughout**
   The frozen-model tau-sweep now evaluates on the *train* partition
   (train_idx) instead of the test partition.  Column names changed from
   test_mi / test_mi_std → train_mi / train_mi_std.  baseline_mi also
   sourced from train_mi.  Consistent with every other mode.

3. **Merge workflow.py into rigorous.py (single file)**
   rigorous.py now contains _find_linear_region, _extrapolate_mi,
   _post_process_and_correct and AnalysisWorkflow directly.  workflow.py
   becomes a thin backward-compat shim that re-exports those names.
   analysis/__init__.py imports AnalysisWorkflow from rigorous.
   Bias-correction WLS regression now fits train_mi (not test_mi) vs
   1/gamma, consistent with every other mode.

4. **Fix dimensionality parallelisation (n_workers > 1 was ignored)**
   Added module-level _run_single_split_task (picklable for Pool.map)
   and _dispatch_splits helper.  Strategy:
   - n_splits > 1 & n_workers > 1 → Pool across splits, each split
     gets n_workers=1 internally (avoids nested pools).
   - n_splits == 1 & n_workers > 1 → n_workers forwarded to the inner
     ParameterSweep for sweep-grid parallelism.
   Both interaction and intrinsic dimensionality benefit.

5. **Update tests & documentation**
   test_workflow_internals.py: imports from rigorous, train_mi column.
   test_precision.py: assert train_mi in df.columns.
   test_run.py: mock DataFrames use train_mi.
   test_visualize.py: rigorous fixture uses train_mi.
   plot.py plot_bias_correction_fit: uses train_mi column.
   results.py: precision plot fallback handles train_mi → mi_mean.
   NEURALMI_REFERENCE.md + docs/: update column names and module refs.
   Tutorials 04 & 06 (ipynb + raw py): update column references.

All 213 tests pass.
